refactor: migrate to workspace
Now the macro crate was only being used as a dependency, but not even going to CI. Fixed this.
The main crate stays rooted, it's idiomatic as long as it's truly main for external consumers
