+++ crates/circuit_component_macro/src/lib.rs
