+++ crates/circuit_component_macro/src/parse_sig.rs
