+++ crates/circuit_component_macro/tests/fail/self_parameter.rs
