+++ crates/circuit_component_macro/tests/fail/too_many_inputs.rs
