+++ crates/circuit_component_macro/tests/integration.rs
