+++ crates/circuit_component_macro/tests/success/no_inputs.rs
