+++ crates/circuit_component_macro/tests/success/one_input.rs
