+++ crates/circuit_component_macro/tests/success/sixteen_inputs.rs
