+++ crates/circuit_component_macro/tests/success/three_inputs.rs
