+++ crates/circuit_component_macro/tests/success/two_inputs.rs
